+++ utils/authTokenApi.ts
@@ -1,10 +1,7 @@
-import { useMutation, useQuery, useQueryClient } from "@tanstack/react-query";
 import axios, { AxiosError } from "axios";
 import auth from "@react-native-firebase/auth";
 import { normalizeTrackResponse } from "./api";
-import { PrivateUserData, ResponseObject, TrackResponse } from "./types";
-import { useAuth, useUser } from "@/hooks";
-import { deleteSecretFromKeychain } from "./keychainStorage";
+import { ResponseObject, TrackResponse } from "./types";
 
 const catalogApi = process.env.EXPO_PUBLIC_WAVLAKE_API_URL;
 const enableResponseLogging = Boolean(
@@ -62,129 +59,6 @@ const requestInterceptor = catalogApiClient.interceptors.request.use(
     return config;
   },
 );
-
-export const usePrivateUserData = (enabled: boolean) => {
-  return useQuery<PrivateUserData>({
-    queryFn: async () => {
-      const { data } = await catalogApiClient
-        .get<ResponseObject<PrivateUserData>>(`/accounts`)
-        .catch((error) => {
-          console.log("usePrivateUserData error", error);
-          throw error;
-        });
-
-      return data.data;
-    },
-    queryKey: ["userData"],
-    enabled,
-  });
-};
-
-interface UserEditForm {
-  name: string;
-  ampSat: string;
-  artwork?: any;
-  artworkUrl?: any;
-  uid: string;
-}
-
-export const useEditUser = () => {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: async (
-      user: Partial<Omit<UserEditForm, "ampSat"> & { ampMsat: number }>,
-    ) => {
-      const formIsEmpty = !Object.values(user).some(
-        (value) => value !== undefined && value !== null,
-      );
-      if (formIsEmpty) {
-        return;
-      }
-
-      const requestFormData = new FormData();
-
-      if (user.name) {
-        requestFormData.append("name", user.name);
-      }
-
-      if (user.ampMsat) {
-        requestFormData.append("ampMsat", user.ampMsat.toString());
-      }
-
-      if (user.artwork) {
-        requestFormData.append("artwork", user.artwork as any);
-      }
-
-      const { data } = await catalogApiClient.put<
-        ResponseObject<{ userId: string }>
-      >("/accounts", requestFormData, {
-        headers: {
-          Accept: "application/json",
-          "Content-Type": "multipart/form-data",
-        },
-        transformRequest: (data) => data, // Prevent axios from trying to transform the FormData
-      });
-
-      return data.data;
-    },
-    onSuccess(data) {
-      queryClient.invalidateQueries({ queryKey: ["userData"] });
-    },
-  });
-};
-
-export const useDeleteUser = () => {
-  const queryClient = useQueryClient();
-  const { logout } = useAuth();
-  const { signOut } = useUser();
-  return useMutation({
-    mutationFn: async () => {
-      const { data } =
-        await catalogApiClient.put<ResponseObject<never>>("/accounts/disable");
-
-      return data;
-    },
-    onSuccess(data) {
-      // nostr logout
-      logout();
-      // firebase logout
-      signOut();
-      // delete user's nostr secret
-      deleteSecretFromKeychain();
-      queryClient.invalidateQueries({ queryKey: ["userData"] });
-    },
-  });
-};
-
-export interface WalletConnection {
-  name: string;
-  pubkey: string;
-  requestMethods: WalletConnectionMethods[];
-  lastUsed?: string;
-  msatBudget: number;
-  maxMsatPaymentAmount: number;
-}
-
-export type WalletConnectionMethods =
-  | "get_balance"
-  | "pay_invoice"
-  | "make_invoice"
-  | "lookup_invoice";
-
-export const useCreateConnection = (onSuccess?: Function) => {
-  return useMutation({
-    mutationFn: async (connection: WalletConnection) => {
-      const { data } = await catalogApiClient.post<
-        ResponseObject<WalletConnection>
-      >(`/accounts/connections`, connection, {});
-      return data.data;
-    },
-    onSuccess(data) {
-      onSuccess?.(data);
-    },
-  });
-};
 
 interface Transaction {
   comment: any;
@@ -267,86 +141,4 @@ export const saveUserIdentity = async (data: {
     ResponseObject<{ userId: string }>
   >(`/accounts/log-identity`, data);
   return response.data;
-};
-
-export const useCreateNewUser = () => {
-  return useMutation({
-    mutationFn: async (body: { username?: string; pubkey: string }) => {
-      const { data } = await catalogApiClient.post<
-        ResponseObject<{
-          username: string;
-          profileUrl: string;
-          pubkey: string;
-          loginToken: string;
-        }>
-      >(`/accounts/user`, body);
-
-      return data.data;
-    },
-  });
-};
-
-// this API endpoint is gaurded by an IP region check
-export const useCreateNewVerifiedUser = () => {
-  return useMutation({
-    mutationFn: async (body: {
-      username?: string;
-      firstName?: string;
-      lastName?: string;
-      pubkey: string;
-    }) => {
-      const { data } = await catalogApiClient.post<
-        ResponseObject<{
-          username: string;
-          profileUrl: string;
-          pubkey: string;
-          loginToken: string;
-        }>
-      >(`/accounts/user/verified`, body);
-
-      return data.data;
-    },
-  });
-};
-
-export const useAccountTracks = () => {
-  return useQuery({
-    queryKey: ["accountTracks"],
-    queryFn: async () => {
-      const { data } =
-        await catalogApiClient.get<ResponseObject<TrackResponse[]>>(
-          `/tracks/account`,
-        );
-
-      return data.data;
-    },
-  });
-};
-
-export const useGetInboxLastRead = () => {
-  const { pubkey } = useAuth();
-  return useQuery({
-    queryKey: ["inboxlastRead"],
-    queryFn: async () => {
-      const { data } = await catalogApiClient.get<ResponseObject<string>>(
-        `/accounts/inbox/lastread`,
-      );
-
-      return data.data;
-    },
-    enabled: Boolean(pubkey),
-    gcTime: 5 * 60 * 1000,
-  });
-};
-
-export const useSetInboxLastRead = () => {
-  return useMutation({
-    mutationFn: async () => {
-      const { data } = await catalogApiClient.put<ResponseObject<never>>(
-        `/accounts/inbox/lastread`,
-      );
-
-      return data;
-    },
-  });
 };
